backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0
PyPDF2==3.0.1
pandas==2.1.4
six==1.16.0
kombu==5.3.4
billiard==4.2.0
vine==5.1.0
dj-database-url==2.1.0